refactor: update tests + splite code to different files
